backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-generativeai>=0.8.0
groq>=0.12.0
httpx>=0.28.0
supabase>=2.11.0
python-dotenv>=1.0.0
pillow>=11.0.0
python-multipart>=0.0.18
